File rename for chemCore base class
